chore: lib refactor
