+++ src/app/(auth)/profile/edit/page.tsx
@@ -1,6 +1,6 @@
 import EditCompanyForm from '@/features/auth/EditCompanyForm';
-import { getAuthSession } from '@/lib/nextAuth/auth';
-import { db } from '@/lib/db';
+import { getAuthSession } from '@/lib/auth';
+import { db } from '@/lib/prisma.db';
 
 export default async function Page() {
   const session = await getAuthSession();
