+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,5 @@
 import NextAuth from 'next-auth/next';
-import { authOptions } from '@/lib/nextAuth/auth';
+import { authOptions } from '@/lib/auth';
 
 const handler = NextAuth(authOptions);
