+++ src/app/api/cart/add/route.ts
@@ -1,5 +1,5 @@
-import { getAuthSession } from '@/lib/nextAuth/auth';
-import { db } from '@/lib/db';
+import { getAuthSession } from '@/lib/auth';
+import { db } from '@/lib/prisma.db';
 
 async function handler(req: Request) {
   const session = await getAuthSession();
