+++ src/app/bar/page.tsx
@@ -1,5 +1,5 @@
-import { getAuthSession } from '@/lib/nextAuth/auth';
-import { db } from '@/lib/db';
+import { getAuthSession } from '@/lib/auth';
+import { db } from '@/lib/prisma.db';
 
 export default async function Page() {
   return <div className='flex flex-col gap-6 mt-12'>BAR</div>;
