+++ src/app/verify-email/page.tsx
@@ -1,6 +1,6 @@
 import VerifyEmail from '@/features/VerifyEmail';
-import { getAuthSession } from '@/lib/nextAuth/auth';
-import { db } from '@/lib/db';
+import { getAuthSession } from '@/lib/auth';
+import { db } from '@/lib/prisma.db';
 
 export default async function Page() {
   const session = await getAuthSession();
