+++ src/app/welcome/page.tsx
@@ -1,4 +1,4 @@
-import { getAuthSession } from '@/lib/nextAuth/auth';
+import { getAuthSession } from '@/lib/auth';
 import Link from 'next/link';
 
 export default async function Welcome() {
