+++ src/features/auth/SignUp.tsx
@@ -5,7 +5,7 @@ import { signIn } from 'next-auth/react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import type { SignUpFormData } from '@/types/auth.form-types';
-import { AuthSignUpValidator } from '@/lib/validators/user-auth';
+import { AuthSignUpValidator } from '@/lib/validator.auth-form';
 
 export default function SignUp() {
   const router = useRouter();
