+++ src/types/auth.form-types.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 import {
   AuthSignInValidator,
   AuthSignUpValidator,
-} from '../lib/validators/user-auth';
+} from '../lib/validator.auth-form';
 
 export type SignInFormData = z.infer<typeof AuthSignInValidator>;
 export type SignUpFormData = z.infer<typeof AuthSignUpValidator>;
